suspense build error fix
